requirements.txt
pytorch-lightning>=1.8.0
hydra-core==1.3
hydra_colorlog==1.2
wandb==0.15
rich==13.6
pandas>=2.0
transformers==4.34
datasets==2.14
spacy==3.7
jsonlines==4.0.0
scipy==1.11.3
sentencepiece==0.1.99
protobuf==3.20
nltk==3.8.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz